requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.13
openai==1.3.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
rich==13.7.0
tiktoken>=0.4.0
streamlit>=1.24.0
chromadb>=0.3.26
